feat: extract acquisition services
