+++ DataAcquisition.Application/Abstractions/IHeartbeatMonitor.cs
@@ -1,0 +1,9 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace DataAcquisition.Application.Abstractions;
+
+public interface IHeartbeatMonitor
+{
+    Task MonitorAsync(DeviceConfig config, CancellationToken ct = default);
+}
