+++ DataAcquisition.Application/Abstractions/IModuleCollector.cs
@@ -1,0 +1,9 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace DataAcquisition.Application.Abstractions;
+
+public interface IModuleCollector
+{
+    Task CollectAsync(DeviceConfig config, Module module, IPlcClientService client, CancellationToken ct = default);
+}
