+++ DataAcquisition.Application/PlcRuntime.cs
@@ -1,0 +1,6 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace DataAcquisition.Application;
+
+public sealed record PlcRuntime(CancellationTokenSource Cts, Task Running);
