--- DataAcquisition.Infrastructure/DataAcquisitions/PlcRuntime.cs
@@ -1,6 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-
-namespace DataAcquisition.Infrastructure.DataAcquisitions;
-
-public sealed record PlcRuntime(CancellationTokenSource Cts, Task Running);
